[expo-updates][Android] Get downloaded update IDs
Why

In order for the dev client to be able to indicate in the UI whether a particular update is available to load offline, we need to add a method to the expo-updates-interface to let it see which updates are stored locally (in sqlite) and ready to load.

It should be sufficient for this method to return a list of UUIDs for updates that have the READY status in sqlite.

How

New method

```java
  void storedUpdateIdsWithConfiguration(HashMap<String, Object> configuration, Context context, UpdateCallback callback);
```

is added to the interface for Android.

The implementation in `UpdatesDevLauncherController` calls into a new method in `DatabaseLauncher`, which uses a new query in `UpdateDao` to retrieve READY update UUIDs.

Test Plan

New unit tests are added to the existing `DatabaseLauncherTest` (Android test class) to verify the changes.